only go to page if we don't have it yet

+++ main.py
@@ -1,3 +1,4 @@
+import os
 from playwright.sync_api import sync_playwright, TimeoutError as PlaywrightTimeoutError
 
 from time import sleep
@@ -28,16 +29,14 @@ def try_goto(link, browser, timeout=1):
     return page
 
 
-def save_to_file(link, page):
+def ensure_file(link, browser):
     filename = f'download/{link[6:]}.html'  # remove the language prefix, add folder and extension
-    with open(filename, 'w') as f:
-        f.write(page.content())
-
-
-def download(link, browser):
+    if os.path.isfile(filename):
+        return
     page = try_goto(f'{base_url}{link}', browser)
     if page is not None:
-        save_to_file(link, page)
+        with open(filename, 'w') as f:
+            f.write(page.content())
         page.close()
     sleep(uniform(0.1, 0.3))  # avoid making calls too close by keeping the page open between 0.1 and 3 seconds    
 
@@ -49,7 +48,7 @@ def read_index(i, browser):
         return
     link_elements = index_page.query_selector_all("article.vehicle-card a.base-a-link")
     for link_element in link_elements:
-        download(link_element.get_attribute("href"), browser)
+        ensure_file(link_element.get_attribute("href"), browser)
     index_page.close()
